requirements.txt
requests>=2.20.0
python-dotenv>=1.0.0
